#25 fix: 소셜로그인 파일 이름 수정
소셜로그인 파일 이름을 FE 규칙에 맞게 수정
