+++ src/pages/login/LoginGoogle.js
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router-dom";
 import { useEffect } from "react";
 import axios from "axios";
 
-const Logingoogle = (props) => {
+const LoginGoogle = (props) => {
     const code = new URL(window.location.href).searchParams.get("code");
     useEffect(() => {
         const googleLogin = async () =>{
@@ -32,3 +32,4 @@ const Logingoogle = (props) => {
         </div>
     )
 }
+export default LoginGoogle;
